+++ packages/core/package.json
@@ -48,7 +48,6 @@
     "express-graphql": "^0.12.0",
     "glob": "^8.1.0",
     "graphql": "^15.3.0",
-    "graphql-http": "^1.22.0",
     "graphql-subscriptions": "^2.0.0",
     "graphql-ws": "^5.14.0",
     "http-terminator": "^3.2.0",
