+++ packages/core/src/indexing-server/service.ts
@@ -1,15 +1,13 @@
 import { makeExecutableSchema } from "@graphql-tools/schema";
-import cors from "cors";
 import express from "express";
 import { graphqlHTTP } from "express-graphql";
 import { PubSub } from "graphql-subscriptions";
 import { useServer } from "graphql-ws/lib/use/ws";
-import { createHttpTerminator } from "http-terminator";
-import { createServer, Server } from "node:http";
 import { Server as WebSocketServer } from "ws";
 
 import { EventStore } from "@/event-store/store";
 import type { Common } from "@/Ponder";
+import { Server } from "@/utils/server";
 
 import {
   FINALITY_CHECKPOINT,
@@ -21,16 +19,13 @@ import {
 } from "./resolvers";
 import { indexingSchema } from "./schema";
 
-// TODO: Refactor common GQL server code with ServerService
 export class IndexingServerService {
   private common: Common;
   private eventStore: EventStore;
   private pubsub: PubSub;
+  private server: Server;
 
-  port: number;
   app?: express.Express;
-
-  private terminate?: () => Promise<void>;
 
   isSyncComplete = false;
 
@@ -43,77 +38,27 @@ export class IndexingServerService {
   }) {
     this.common = common;
     this.eventStore = eventStore;
-    this.port = this.common.options.indexingPort;
+
+    this.server = new Server({
+      common,
+      port: common.options.indexingPort,
+    });
 
     // https://www.apollographql.com/docs/apollo-server/data/subscriptions#the-pubsub-class
     this.pubsub = new PubSub();
   }
 
-  async start() {
-    this.app = express();
-    this.app.use(cors());
+  get port() {
+    return this.server.port;
+  }
 
-    // TODO: Register metrics for IndexingServerService similar to ServerService
+  async start() {
+    const server = await this.server.start();
+    this.common.metrics.ponder_server_port.set(this.server.port);
 
     const schema = makeExecutableSchema({
       typeDefs: indexingSchema,
       resolvers: getResolvers(this.eventStore, this.pubsub),
-    });
-
-    const server = await new Promise<Server>((resolve, reject) => {
-      const server = createServer(this.app)
-        .on("error", (error) => {
-          if ((error as any).code === "EADDRINUSE") {
-            this.common.logger.warn({
-              service: "indexing-server",
-              msg: `Port ${this.port} was in use, trying port ${this.port + 1}`,
-            });
-            this.port += 1;
-            setTimeout(() => {
-              server.close();
-              server.listen(this.port);
-            }, 5);
-          } else {
-            reject(error);
-          }
-        })
-        .on("listening", () => {
-          // TODO: Set metrics for indexing server port
-          resolve(server);
-        })
-        .listen(this.port);
-    });
-
-    const terminator = createHttpTerminator({ server });
-    this.terminate = () => terminator.terminate();
-
-    this.common.logger.info({
-      service: "indexing-server",
-      msg: `Started listening on port ${this.port}`,
-    });
-
-    // TODO: Add server request handlers for metrics
-
-    // Server will respond as unhealthy until historical events have
-    // been processed OR 4.5 minutes have passed since the app was created.
-    // Similar to implementation in ServerService
-    this.app.get("/health", (_, res) => {
-      if (this.isSyncComplete) {
-        return res.status(200).send();
-      }
-
-      const max = this.common.options.maxHealthcheckDuration;
-      const elapsed = Math.floor(process.uptime());
-
-      if (elapsed > max) {
-        this.common.logger.warn({
-          service: "indexing-server",
-          msg: `Historical sync duration has exceeded the max healthcheck duration of ${max} seconds (current: ${elapsed}). Sevice is now responding as healthy and may serve incomplete data.`,
-        });
-        return res.status(200).send();
-      }
-
-      return res.status(503).send();
     });
 
     const graphqlMiddleware = graphqlHTTP({
@@ -121,7 +66,7 @@ export class IndexingServerService {
       graphiql: true,
     });
 
-    this.app.use("/graphql", graphqlMiddleware);
+    this.server.app?.use("/graphql", graphqlMiddleware);
 
     // create and use the websocket server
     const wsServer = new WebSocketServer({
@@ -133,15 +78,12 @@ export class IndexingServerService {
   }
 
   async kill() {
-    await this.terminate?.();
-    this.common.logger.debug({
-      service: "indexing-server",
-      msg: `Stopped listening on port ${this.port}`,
-    });
+    await this.server.kill();
   }
 
   setIsSyncComplete() {
     this.isSyncComplete = true;
+    this.server.isHealthy = true;
 
     this.common.logger.info({
       service: "indexing-server",
