+++ packages/core/src/utils/server.ts
@@ -1,0 +1,110 @@
+import cors from "cors";
+import express from "express";
+import { createHttpTerminator } from "http-terminator";
+import { type Server as ServerInterface, createServer } from "node:http";
+
+import type { Common } from "@/Ponder";
+
+export class Server {
+  private common: Common;
+
+  port: number;
+  app?: express.Express;
+
+  private terminate?: () => Promise<void>;
+
+  constructor({ common, port }: { common: Common; port: number }) {
+    this.common = common;
+    this.port = port;
+  }
+
+  isHealthy = false;
+
+  async start() {
+    this.app = express();
+    this.app.use(cors());
+
+    const server = await new Promise<ServerInterface>((resolve, reject) => {
+      const server = createServer(this.app)
+        .on("error", (error) => {
+          if ((error as any).code === "EADDRINUSE") {
+            this.common.logger.warn({
+              service: "server",
+              msg: `Port ${this.port} was in use, trying port ${this.port + 1}`,
+            });
+            this.port += 1;
+            setTimeout(() => {
+              server.close();
+              server.listen(this.port);
+            }, 5);
+          } else {
+            reject(error);
+          }
+        })
+        .on("listening", () => {
+          resolve(server);
+        })
+        .listen(this.port);
+    });
+
+    const terminator = createHttpTerminator({ server });
+    this.terminate = () => terminator.terminate();
+
+    this.common.logger.info({
+      service: "server",
+      msg: `Started listening on port ${this.port}`,
+    });
+
+    this.app.post("/metrics", async (_, res) => {
+      try {
+        res.set("Content-Type", "text/plain; version=0.0.4; charset=utf-8");
+        res.end(await this.common.metrics.getMetrics());
+      } catch (error) {
+        res.status(500).end(error);
+      }
+    });
+
+    this.app.get("/metrics", async (_, res) => {
+      try {
+        res.set("Content-Type", "text/plain; version=0.0.4; charset=utf-8");
+        res.end(await this.common.metrics.getMetrics());
+      } catch (error) {
+        res.status(500).end(error);
+      }
+    });
+
+    // By default, the server will respond as unhealthy until historical events have
+    // been processed OR 4.5 minutes have passed since the app was created. This
+    // enables zero-downtime deployments on PaaS platforms like Railway and Render.
+    // Also see https://github.com/0xOlias/ponder/issues/24
+    this.app.get("/health", (_, res) => {
+      if (this.isHealthy) {
+        return res.status(200).send();
+      }
+
+      const max = this.common.options.maxHealthcheckDuration;
+      const elapsed = Math.floor(process.uptime());
+
+      if (elapsed > max) {
+        this.common.logger.warn({
+          service: "server",
+          msg: `Historical sync duration has exceeded the max healthcheck duration of ${max} seconds (current: ${elapsed}). Sevice is now responding as healthy and may serve incomplete data.`,
+        });
+        return res.status(200).send();
+      }
+
+      return res.status(503).send();
+    });
+
+    return server;
+  }
+
+  async kill() {
+    await this.terminate?.();
+
+    this.common.logger.debug({
+      service: "server",
+      msg: `Stopped listening on port ${this.port}`,
+    });
+  }
+}
